+++ src/common.ts
@@ -38,9 +38,6 @@ export interface UploadDisplayDragEvents {
 
 export type ImageEvent<T = any> = SyntheticEvent<T>;
 
-/**
- * 通用全局类型
- * */
 export type PlainObject = { [key: string]: any };
 
 export type OptionData = {
@@ -60,6 +57,9 @@ export type TreeOptionData<T = string | number> = {
   content?: string | TNode;
 } & PlainObject;
 
+/**
+ * 通用全局类型
+ * */
 export type SizeEnum = 'small' | 'medium' | 'large';
 
 export type ShapeEnum = 'circle' | 'round';